Pass layout to bespoke irgen helpers
Summary:
Rather than pulling the layout from the array's type in each place, we pass it into translateDispatchBespoke.

We may lose a small amount of information here, but I don't think it's that likely based on how we define this layout. The layout is pulled from the array type (for live + optimizing) and is BespokeTop (for the diamond) - hard to improve on either. If we do get a tighter layout, we'll improve code in simplify or irlower (coming soon).

Differential Revision: D24926557

fbshipit-source-id: 84db9ac916e0c3fd943a0745ec1f4c64771be4f3