Match APIs for AppendMove and GetPosKey / GetPosVal
Summary:
In order to support fully-layout-generic code, we want a function to call for BespokeAppend with the right PRc/CRc semantics. We'll need to call this function on:
- A generic ArrayData
- A vanilla ArrayData of a known type (so use PackedArray, MixedArray, or SetArray versions)
- A bespoke ArrayData of unknown layout (so use BespokeArray version)
- A bespoke ArrayData of known layout (so use layout->fnAppendMove)

We also rename GetKey / GetVal to GetPosKey / GetPosVal. The latter names are both clearer and consistent with the methods on the vanilla arrays and on BespokeArray.

Differential Revision: D24935796

fbshipit-source-id: e12ee941d3eba9db2ce9772988b05978718d2df0