Do not use Decl.make_env for deferred decl
Summary:
When the direct decl parser is disabled, Decl.make_env fetches a file AST from Ast_provider (i.e., the ParserHeap), declares everything in that AST, and writes those decls to shared memory. When the direct decl parser is enabled, Decl.make_env fetches a file's source text from File_provider (i.e., from disk or from the FileHeap), direct-decl-parses it, writes the resulting decls to shared memory or a local decl cache (depending on the selected provider backend), computes folded class declarations, and writes them to shared memory as well.

The pre-direct-decl-parser behavior works just fine. We do extra work if another worker races to compute a file's decls, but not much, since file ASTs are cached in the ParserHeap. The behavior with the direct decl parser, however, is a problem! In a full init, we'll have fully populated our shallow decl heaps before typechecking begins, so performing a direct-decl-parse-and-cache in response to a deferred decl is completely wasted work!

Reviewed By: shiqicao

Differential Revision: D24937773

fbshipit-source-id: 50872fd173e143faabd704555666047923632a52