Match OCaml typing reasons on Taccess hints
Summary:
In OCaml, we flatten nested type-constant-path typehints into an Haccess node with a list of constant names that appeared in the path. We used to represent Taccess similarly, but stopped doing so in {D24620323 (https://github.com/facebook/hhvm/commit/4e585b238046144ba9fa4f7563d452702d622fbb)}. Since Haccess is still represented with a flattened list, and doesn't contain enough position information to construct accurate positions for each Taccess projection, each Taccess is given a position referring to the entire path.

The direct decl parser does have this position information, and currently uses it. This might be a better behavior in the long run, but it's an obstacle to ensuring that the direct decl parser behaves the same as OCaml decl in the short term. For now, this diff rewrites these positions to be the same as OCaml decl.

Reviewed By: shiqicao

Differential Revision: D24939218

fbshipit-source-id: 14b58de33781e21fef234de3fe222fbb5876a67d