Make Bespoke IR ops fully-generic
Summary:
In order to emit fully-layout-generic code, we need to have IR ops that can handle both vanilla and bespoke array-likes. We modify Bespoke... to be those ops.

To do so, we drop the BespokeLayout parameter on these ops and re-implement their irlower method as a complex dispatch to one of:
- a concrete BespokeLayout's method
- BespokeArray's method, generic over all bespokes
- a concrete vanilla array-like's method (Packed, Mixed, Set)
- ArrayData itself, or some other fully-generic helper

This logic is repetitive but uses lots of strongly-typed functions. I see no good alternative to doing it by macro. Finally, we have two edge cases: BespokeElem and BespokeEscalateToVanilla. BespokeElem doesn't have the regular dispatch as above; I hand-coded various specializations that already exist in minstr-helpers, and fell back to a relatively generic helper. BespokeEscalateToVanilla is the only IR op that is truly bespoke-only.

Differential Revision: D24943704

fbshipit-source-id: 7da6a4cba5689b6170bd11d5cb812f884f9fd465