Fix bug in BespokeElemD
Summary: We've changed the contract for BespokeElem slightly: if the Elem targets a ClsMeth, then caller is allowed to escalate it. We can drop the ClsMeth behavior in member-instructions.h now. In fact, it's incorrect: because Elem escalates the array, the first call to BespokeElem there could convert it to a vanilla array, at which point the second call would break.

Reviewed By: ricklavoie

Differential Revision: D24974615

fbshipit-source-id: 07e6a59bb01159085a045df8f8cb57694cf5b92a