emit_bindaddr_stub: do not use RIP relative addressing for addrs outside of TC
Summary:
emit_bindaddr_stub() checked semi-correctly whether the smashable address is
RIP relative addressable from the current frontier, but then we allocated stub
space using allocTCStub(), which may return a previously used stub that has
nothing to do with the frontier and may be too far away from the addr.

Instead, use RIP relative addressing only for smashable addresses within the
data segment, which need it for relocatability and which are guaranteed to be
within 2GB from the code. Do not try to be smart about addresses in general
heap.

Fixes random memory corruption issues caused by smashing addresses 2^32 bytes
lower than what was supposed to be smashed and segfaults when we were lucky and
we tried to smash a small portion of unmapped low memory address space.

Also, add asserts so that similar issues are detected earlier in the future.

Reviewed By: ricklavoie

Differential Revision: D24982905

fbshipit-source-id: bdb260d3fe4a07314673698a5e67717ac0f84cc4