defer decl at memory_mb_threshold
Summary:
The previous diff had a huge summary about the overview and motives for introducing "defer if we're asked to decl but have exceeded memory limit".

This diff implements that change.

This diff doesn't yet actually turn anything on. It merely supports the flag in hh.conf. I'll leave it to future work to actually enable the feature.

Reviewed By: CatherineGasnier

Differential Revision: D24978115

fbshipit-source-id: 1825d9ec1c827d49a8ddc7c747035287f49d66d3