tco_max_typechecker_worker_memory_mb
Summary:
This is a small bit of boilerplate. Most of the settings that control typechecker performance (num_local_workers, parallel_type_checking_thresholdd, defer_class_ddeclaration_threshold) are inside tcopt.

The only one that's not is max_typechecker_worker_memory_mb. It lives solely in genv, and has to be passed down as a parameter `~memory_cap` into each of the functions in typing_check_service.ml.

For uniformity, I'm going to move this last remaining setting into tcopt.

This diff merely introduces the tcopt flag. The next diff will switch over to it.

Reviewed By: CatherineGasnier

Differential Revision: D24977673

fbshipit-source-id: 06340b416c01e941b0f8bc94c5187a9f2db3c7ee