get memory_cap from tcopt
Summary:
This is a small bit of boilerplate. Most of the settings that control typechecker performance (num_local_workers, parallel_type_checking_thresholdd, defer_class_ddeclaration_threshold) are inside tcopt.

The only one that's not is max_typechecker_worker_memory_mb. It lives solely in genv, and has to be passed down as a parameter `~memory_cap` into each of the functions in typing_check_service.ml.

For uniformity, I'm going to move this last remaining setting into tcopt.

The previous diff introduced the tcopt flag for it. This diff switches over from memory_cap to the tcopt parameter.

Reviewed By: CatherineGasnier

Differential Revision: D24977712

fbshipit-source-id: 827bea1874067fc081bce972e0e4ca5876999f79