Add ArrayLayout, generalizing BespokeLayout and vanilla
Summary:
ArrayLayout is basically a little piece of ArraySpec that we've chosen to extract because it's useful in its own right. It's the "m_sort" piece, which captures all bespoke::Layouts, Top, Vanilla, and Bottom. By pulling this piece out, we have a safer, JIT-only abstraction that allows us to generate code for arbitrary array-likes at arbitrary granularities.

The result of bespoke layout selection will be an ArraySpec - NOT a bespoke::Layout! - because we may choose vanilla as a source type, or "specialize" a sink on Top.

Differential Revision: D25015244

fbshipit-source-id: 385de51576d79da24b4dbfccbba32f884623adcd