Fix hashbang parsing for .hack and .hackpartial files, disallow #! style comments everywhere else
Summary:
Our hashbang parsing behavior is pretty incorrect for `.hack` and `.hackpartial` files: specifically, hashbangs get parsed as regular comments in those files and trivia instead of as a hashbang, since our hashbang token syntax is only implemented for the markup case. Since we need to disable `#!` style comments to fully remove `#` style comments (keeping one but not the other would be confusing for developers and also bad for syntax highlighters), both these bugs need to be fixed.

This diff fixes the bug by doing the following:

- MarkupSection consists of two parts:
markup_hashbang and markup_suffix.

- If there's a hashbang and markup, parse correctly, behavior is unchanged
- If there's a hashbang with no markup, parse with a missing markup_suffix field (and throw errors if appropriate
- If there's no hashbang, behavior is unchanged from before

This requires changing:
- the lexer to correctly parse hashbangs when there isn't a `<?hh`
- the declaration parser to not return None when there's no `<?hh`
- the mode parser to correctly handle hashbangs with no `<?hh`
- finally, the lexer again to disallow #! comments.

Reviewed By: shayne-fletcher

Differential Revision: D24870769

fbshipit-source-id: 32e1dc56bf36dc5db43e806ed707d6da11531a25