Rewrite ArraySpec in terms of ArrayLayout
Summary:
In this diff, we delete BespokeLayout and replace all of its usages for ArrayLayout. Because ArrayLayout also includes Vanilla and Top, ALL ArraySpecs have an ArrayLayout, so the mapping between old and new APIs requires a tiny bit of work. Say we have an ArraySpec "spec":
| old API | new API
| bool(spec.bespokeLayout()) | spec.bespoke()
| spec.bespokeLayout() | spec.layout().bespokeLayout()
| spec.narrowToVanilla() | spec.narrowToLayout(ArrayLayout::Vanilla())
| spec.narrowToBespokeLayout(layout) | spec.narrowToLayout(ArrayLayout(layout.index())

etc. Aside from these small "syntactic" differences, the two APIs are basically the same, but one place where our code gets a lot simpler is the type operations on ArraySpec (because all layout-related logic is now handled by ArrayLayout).

Differential Revision: D25015322

fbshipit-source-id: 566dd7f92a47b966f7b5a096f8d6bc3a3cefa373