Topological ordering optimizations for bespoke types
Summary:
To avoid having to do graph traversal to compute join and meet for bespoke types, instead cache their ancestors and descendants in topological order so that they can be computed by a simple parallel walk.

In doing so, drop the requirement that LayoutIndices must themselves form a topological order.

Also correct my inability to spell.

Differential Revision: D24995189

fbshipit-source-id: 979703a74b3d4c75c6c033e16dd117be2dee85c5