basic handling of dynamic
Summary:
In trying to experiment with making IFC more robust, I'm starting trying to handle dynamic.

This diff adds basic support for reads and writes to dynamics. Dynamics have a single invariant policy, so their subtyping rules are identical to `Tgeneric`.

Using a dynamic as a prim, is pretty straightforward, the resulting prim gets the same policy as the dynamic.

Using a dynamic as an object uses the dynamic's policy as both the self and lump policies. This is technically unsafe because the property we are accessing could have a different policy from the lump, so we emit an error to reflect this.

Reviewed By: mpu

Differential Revision: D24137552

fbshipit-source-id: b0dd891a8b51117f673e3bbb4239922ecc65db10