upload saved state manifold path to remote worker
Summary: Currently the remote workers receive an `init_payload` consisted of the typechecker options and a path containing the naming table. This diff passes an additional piece of information, namely the naming table saved state manifold path, to `init_payload`, so that during initialization the remote workers can download the naming table from this manifold path (implemented in the next diff).

Reviewed By: yinghuitan

Differential Revision: D24907729

fbshipit-source-id: 5550a090a3f46a48f3d5bc81eb88b34cd2f5dce3