Filter out PHPStdLib methods and properties
Summary: In naming, when the `deregister_php_stdlib` parser option is set, we filter out functions, classes, class properties, and methods which have the <<__PHPStdLib>> attribute in .hhi files. When using the direct decl parser, we need to behave the same way. We currently filter out functions and classes with this attribute after performing a direct-decl-parse, but we don't filter out methods and properties yet, because it wasn't possible to tell from a shallow method or property whether this attribute was present before the previous diff in this stack. This change adds this method and property filtering after decl-parsing.

Reviewed By: Wilfred

Differential Revision: D24966500

fbshipit-source-id: 22e3f26de21ef21bab69d47b8ecfd15b2ae57ebf