Add php_std_lib flag to shallow methods and properties
Summary: In naming, when the `deregister_php_stdlib` parser option is set, we filter out functions, classes, class properties, and methods which have the `<<__PHPStdLib>>` attribute in .hhi files. When using the direct decl parser, we will need to behave the same way. We currently filter out functions and classes with this attribute after performing a direct-decl-parse. We don't filter out methods and properties yet, because it isn't currently possible to tell from a shallow method or property whether this attribute was present. This change adds a flag for the __PHPStdLib, making it possible to do this filtering after decl-parsing.

Reviewed By: Wilfred

Differential Revision: D24966497

fbshipit-source-id: 003a7330c5149dd6074e88201ead3ded256d898c