Tighten hierarchy assertions
Summary: I'm pretty sure that ArrayLayout will always handle the case when either of the two layouts is a "basic" layout. We no longer need bespoke::Layout type-ops pre-FinalizeHierarchy. However, this constraint is a bonus on top of the previous diff and not required for correctness, so I've split it out.

Differential Revision: D25015566

fbshipit-source-id: 398cc0030952d6987783e50b8ea1e1cae5e0c8b9