Type checks for bespoke hierarchy
Summary: We implement type checks for the bespoke hierarchy. The initial algorithm is simple. It attempts to find a single mask and compare that tests for all of the concrete descendants of a given node (while rejecting all of the non-descendants), returning it if one exists. Otherwise, if the node has two children (each with single tests), it uses those two tests. If neither strategy works, we fail. All type check tests are precomputed when the hierarchy is finalized.

Differential Revision: D24953937

fbshipit-source-id: b75c4dddaedea97f1a1dffe659a17c8eb09270f3