Change capability to be a variant in Tfun
Summary: Instead of eagerly expanding a missing context on a function header into `HH\Contexts\defaults`, this uses a sentinel variant to note that the function has the default capability. This variant should only be used for top level functions and function type hints, not lambdas as their default contexts come from the local environment.

Reviewed By: losvald

Differential Revision: D25035736

fbshipit-source-id: 1e25b8294f56245d4277d7e7ed3cdc51775e6d95