Introduce __systemlib_reified_generics() special function instead of variable variables
Summary:
We want systemlib to be able to access the reified type structure, which is stored in a local variable called `$0ReifiedGenerics`.

systemlib previously used variable variables for this, which we want to ban. Instead, introduce a `\__systemlib_reified_generics()` that gets transformed to `$0ReifiedGenerics` inside systemlib.

Outside of systemlib nothing is transformed, so userland code cannot use it.

Reviewed By: oulgen

Differential Revision: D24698511

fbshipit-source-id: 826d577aed85ad52695693c4ad697061749f0fc8