Support static string key escalation in MonotypeDict
Summary:
Before, to look up a non-static string in a MonotypeDict<StaticString,T>, we did an expensive "coerce to static" call (lookupStaticString) that has to look at a global, synchronized hash table.

However, the layout of MonotypeDict<StaticString,T> puns the layout of MonotypeDict<String,T> perfectly, so if we relax conditions on checkInvariants, we can just call the MonotypeDict<String,T> methods! There is only a small amount of additional work we must do, but some of it is hilarious. One thing is that when we set a non-static-string key, we have to update the layout to include the type, because the layout may still be the static string one. The other is that there's a test - which includes a mix of static and non-static strings - that we raise when we insert the 2000th element into a particular hash bucket. We have to implement this behavior for MonotypeDict.

Differential Revision: D25067722

fbshipit-source-id: 56fca7f4f2ea3ec7dbc68d646621cf176807072f