Add internal nodes for MonotypeVec and MonotypeDict layouts
Summary:
We'd like to have bespoke::Layout nodes for "monotype vec or empty of a given datatype" and "any monotype vec or empty", so that we can JIT better code for these levels at certain sinks.

In order to make this work we have to modify how we compute the "compare one layout, then mask and compare the rest" - there are layout like EmptyOrMonotypeVec<String> that have three children (in this case: EmptyMonotypeVec, MonotypeVec<String>, and MonotypeVec<PersistentString>), so the existing logic based on "two children with one MaskAndCompareEach" doesn't work. I've replaced the logic with logic that explicitly includes one of a handful of "singleton layouts" - we'll just add EmptyMonotypeVec and EmptyMonotypeDict there and we should be good.

Differential Revision: D25032544

fbshipit-source-id: 87b743c2dda530f08f56589afb1e752ebf5b1c14