Separate rewriting from emitting
Summary: Today AST rewrites occur inline during bytecode emission. The intent is to tease these two phases apart. This diff gets the ball rolling by lifting closure conversion out into a rewriting step that runs before the emit step instead of during. It's in some sense a minor change but it arranges a framework for the work to come.

Reviewed By: shiqicao

Differential Revision: D25058179

fbshipit-source-id: 34b018215263ce9cac40a8607cbd40e2d93d787d