Fix potential write heap buffer overflow in libgd
Summary:
Fuzzer found a case that if a size between -1 and -8 is passed to one of these
functions, we would try to allocate less than 8 bytes and then write this
allocation size in the first 8 bytes, part of which will write after the
allocated memory.

This is not exploitable, as memory allocators usually don't allocate less than
8 bytes. Even if, it would overwrite the extra memory with zeroes on little
endian platforms.

Fix this by always checking the size to be within the expected range (it is
unsigned, so checking for upper bound is sufficient).

Reviewed By: ottoni

Differential Revision: D25029772

fbshipit-source-id: e265b159ee12ea94fbddbd95d5203cf747c9bd96