Change Eval.ForbidDynamicCallsWithAttr to prepare for deprecating the option
Summary:
We intend to remove `hhvm.forbid_dynamic_calls_with_attr` in the future and have the `hhvm.forbid_dynamic_calls_to_*` settings control the notices for raw dynamic calls directly. That is:
- if the level is set to 1, then we only raise notices if the function is not annotated with `<<__DynamicallyCallable>>`
- if the level is set to 2, then we throw if the function is not annotated with `<<__DynamicallyCallable>>`, but we also raise notices if the callsite is not using an explicit function pointer to make the dynamic call
- if the level is set to 3, then we throw for either condition

In this diff, we change the default for `hhvm.forbid_dynamic_calls_with_attr` to true, and it no longer has an effect if the corresponding level for `hhvm.forbid_dynamic_calls_to_func`, etc. is 1 (or 0).

Reviewed By: paulbiss

Differential Revision: D24877373

fbshipit-source-id: d9632c52ef8785ffd0864f2a36448aac353f2c24