Change default capability to be `defaults` instead of bottom
Summary:
This is technically a revert of {D24899446} which was a revert of {D24132229 (https://github.com/facebook/hhvm/commit/735fc5191f51e678542d0ad7179bfd3b9735a072)}, but because of the variant introduced in the previous diff, has a much smaller impact on performance.

The `HH\Contexts\defaults` alias is expanded only at the point of subtyping the capability from context with a function's required capability (during a call) and when comparing the capabilities of two functions. It will also need to be expanded

The key here is that we're no longer storing the literal decl type `Tapply "\HH\Contexts\defaults"` in the decl function type record.

Reviewed By: losvald

Differential Revision: D24899446

fbshipit-source-id: 1d545f4a2b483654e73928329c8294ccd2a12a95