Replace isInitialized checks in systemlib with isNull
Summary:
Checking for NullUninit is in correct since for native functions if the default argument is omitted, we pass Nulluninit. But if the function passes null and there's a check for NullUninit, the behavior changes. This is broken.
```
stream_context_set_option($context, $options);
stream_context_set_option($context, $options, null);
```
should behave the same in order to merge FCall and FCallBuiltin.

Reviewed By: ricklavoie

Differential Revision: D25104212

fbshipit-source-id: d096aecc786c0a49a0b7265ea65c61b799063329