Redirect the callers of functions that failed to be optimized to the fcall helper stub
Summary:
If the JIT failed to emit the prologue of an optimized function, which can
happen if the TC fills up, then the callers of that function wouldn't be smashed
and, instead, would continue to call the profiling prologue.  This would then
cause crashes once code.prof was freed.  This diff simply smashes the callers of
such functions to call the fcall helper stub instead, so that they stop calling
the profiling code.

Reviewed By: jano

Differential Revision: D25141693

fbshipit-source-id: b37fc83ea623928a4cd3391d5c6eb8327c614dcd