Fix a race condition for marking functions as being profiled
Summary:
D24371160 (https://github.com/facebook/hhvm/commit/65e94814a0975b5b22befc5ea01cdcbdf8a8b07a) introduced a race that could cause the set of functions being profiled to be inconsistent. The set of profiled functions was represented both with an AtomicLowPtr vector in ProfData and an atomic flag in the Func. It was possible for a thread to set the atomic flag and, before it inserted the Func in the vector, other threads would see the atomic flag set and continue to execute, potentially iterating through all the profiled functions in the vector, which would still be missing some functions that already had the atomic flag set. The entries for such functions would be null in the vector, which the code suspiciously skipped over.

This diff replaces the AtomicLowPtr vector with an AtomicHashMap, and gets rid of the "Profiling" atomic flag in the Func. By having a single representation of the profiled functions, this avoids the bug described above (similarly to how things worked prior to D24371160 (https://github.com/facebook/hhvm/commit/65e94814a0975b5b22befc5ea01cdcbdf8a8b07a)). This diff also replaces the checks that skipped over null functions with asserts that they're not null.

Reviewed By: jano

Differential Revision: D25077393

fbshipit-source-id: 23c008214951fad9c2729edf2a77a23d9750517b