JIT MonotypeVec getters
Summary:
JIT the getters instead of dispatching to native helpers for MonotypeVecs.

The LdMonotypeVecElem opcode is introduced, which loads the type byte from the layout index and the 8-byte value from the MonotypeVec's data. The layout index load is eliminated if we know the monotype of the MonotypeVec.

For the time being, I have placed the irlower methods for MonotypeVec into irlower-bespoke. We can certainly choose to split these out if we prefer.

Differential Revision: D25129277

fbshipit-source-id: 39a8e63c9663646f25680df51bb08528f8cfe2b5