Fix flaky test/slow/async/resume_throw_hook_timeout.php
Summary:
Depending on where the timeout occurs, the fatal either reports the location or
it does not. The `%s` was intended to handle this optionality, but apparently
it does not match an empty string, which `%S` does, so let's use that.

Not sure why this test used to work until recently though. I remember observing
both situations locally when I added the test and `%s` seemed to work.

Reviewed By: ottoni

Differential Revision: D25158058

fbshipit-source-id: 7523456e847b3ff00d1b7aa16f7ee1ce9da1a1e4