Use the correct pointer in bespoke's freeStaticArray()
Summary:
This pointer was 8 bytes off if string key table was used, as it did not
contain the corresponding rounding logic of the Hashtable::staticAlloc().

This logic was usually triggered by retranslateAll(), which tried to use it for
the same underlying static array and thus raced, freeing the unused array. Then,
jemalloc in non-dbg mode would notice that the array is not right when trying
to recycle it in a background thread, causing random segfaults.

Reviewed By: oulgen

Differential Revision: D25167091

fbshipit-source-id: b76332177cb20692ea7feddd0a95df760b41e201