Layout selection for arbitrary static arrays
Summary:
maybeMonoify already does 90% of what we need to make arbitrary static array constructors bespoke. We expose it and hook into it in layout selection. One problem we run into is that we can't allocate request-local memory in the RTA thread. We solve that by modifying MonotypeDict's and MonotypeVec's MakeReserve to directly produce a static allocation. Since this case is cold, we make it a template argument rather than a regular argument.

Since we're already touching maybeMonoify, we fix a small bug: we should only cache a static array there if it's a static constructor, which is true iff it has staticLoggingArray set.

Differential Revision: D25180548

fbshipit-source-id: 4329f1b087386f3c0b862e5fd61e67ffb67a48e1