Add trivial layout selection algorithm
Summary:
We believe there are a few flaws causing us to encounter large regressions when we enable layout selection:
- If we disable jumpstart, we see a large memory regression because our profiles are huge.
- Since we don't ser/de bespoke profiles, we produce different code pre- and post-jumpstart and vasm-block-counts don't apply.
- Our specialized iterator codegen still assumes that arrays are vanilla, so we'll side-exit any specialized loops with bespoke inputs.

We can work around all of these problems with trivial layout selection algorithms that only use vanilla array-likes. We should be able to specialize sources on vanilla and sinks on either {vanilla,top} - if we make sinks vanilla, we should see our old 0.5% perf hit, and if we make sinks top, we should see an additional instructions penalty.

Differential Revision: D25199253

fbshipit-source-id: ed1cc54bafbbac217d084bec1550ac01475028b7