more shape reads via Shapes::idx
Summary:
This diff enables a bunch more shape read operations for shapes with optional fields and open shapes. The only way to read these fields is with `Shapes::idx`.

I order to track the policies of unnamed fields in an open shape, I added a mixed ptype for open shapes. The information flows therefore just fall out from mixed suptyping.

For subtype calls `Tshape s1 <: Tshape s2` where `s2` is an open shape, we make all the fields in `s1` that are not present in `s2` a subtype of `s2`'s magic extra field.

Reviewed By: mpu

Differential Revision: D24690552

fbshipit-source-id: b5c0cc3346eb6422ed419c137318417478e873c0