Tighten typing
Summary:
It turns out that after D25058179 (https://github.com/facebook/hhvm/commit/20227a17d6d06cd43d850d043ad73076fa7e978d), `emit_program` no longer requires a mutable reference to the AST. That is, there are no  in-place mutations to the AST that happen in that function. That might be a surprise, I'm not sure.

In any case even if there are still changes to come as a result of looking to separate rewrites from emitting, it seems correct to tighten the typing to reflect this now apparent invariant.

Reviewed By: shiqicao

Differential Revision: D25133504

fbshipit-source-id: 04f6519823e6b40fba5dba1c76057dd130919ca0