Jumpstart support for bespoke arrays
Summary:
There's one place today where we serialize static arrays: in order to serialize a jit::Type. However, we don't track the array's layout when we do so; on the other side, the deserialized array is always vanilla.

Add support for round-tripping monotype arrays through serialization. We don't support round-tripping LoggingArrays; these types are rare and unimportant (anything we can do with them we can do, better, with a vanilla array). We don't need to match layout for correctness: we never guard on logging arrays, so the only way one will get into our profile is if it's stored into a TypeProfile (in which case we'd prefer to change it).

Differential Revision: D25232088

fbshipit-source-id: a28754d67a41f186e8498a360091b98ef734064c