Fix bug in serialization logic
Summary:
In order to avoid serializing the same JIT object twice, most of our data structures have an m_serialized bit along with a function "serialize()". This function returns true and sets the bit if we should serialize the object; it returns false if the object is already serialized. We use Class::serialize in write_class() - if !serialize(), we simply dump the raw pointer to our data; else, we dump all the info we need to reconstruct it. Similarly, we use Func::serialize in write_func, Record::serialize in write_record, and Unit::serialize in write_unit.

However, we've misused this API by calling Class::serialize and Func::serialize from the serialization routines for OTHER TYPES. When we do so, we break serialization of that class or func unless it's already been serialized. Because of the order in which we do things, we happen to avoid triggering this bug in prod.

Fix this bug by calling "wasSerialized()" from other types; serialize() should only be called for the serialization routing for its own type.

Differential Revision: D25237812

fbshipit-source-id: 95aaf51d3aacc2f513990ca22861137adc546134