Allow calls to more than just named functions
Summary:
Currently, we only allow named calls. `foo(...)` and `Foo::bar(...)`

However, since some examples extend beyond named calls, it is necessary to allow calling of more than just named calls. For instance: `$x()` and `(() ==> { ... })()` and `__splice__(...)()`

This diff allows calling additional constructs that are not named functions. Thus, if an expression is currently valid in an Expression Tree, it can be invoked as a function.

```
Example`$x()`;
```

is not virtualized, and is typechecked as is. It is desugared at runtime into:

```
(Example $v) ==> {
  return $v->call(
    ...,
    $v->localVar(..., '$x'),
    vec[],
  );
}
```

Note: We still do not allow calling instance methods `$obj->get(...)` as the `ObjGet` node is still not allowed in Expression Trees.

Reviewed By: Wilfred

Differential Revision: D24953541

fbshipit-source-id: 0f9bcc6bd01c15cb5b07ee4a75b0e7fc1f99b5ae