Add ternary operator to core expressions
Summary:
We have been using ternary operators in some examples, so let's allow it in Expression Trees

Unlike other operators (such as the unary and binary operators, ternary is going to be desugared, not virtualized.

Since there is an implicit boolean conversion happening, we will want to virtualize the boolean coercion, much like we do with if/while/for.

Thus:

```
Example`$x ? $y : $z`;
```

gets virtualized to

```
`$x->__bool() ? $y : $z`;
```

And desugared to

```
(Example $v) ==> {
  return $v->ternary(
    ...
    $v->methCall(
      ...,
      $v->localVar(..., '$x'),
      '__bool',
      vec[],
    ),
    $v->localVar(..., '$y'),
    $v->localVar(..., '$z'),
  );
}
```

Reviewed By: Wilfred

Differential Revision: D24936267

fbshipit-source-id: 58a7ffbde3ac209a1474e0612168bd13b6db9d18