Implement xor/and/cmp bespoke type test algorithm
Summary:
We introduce a new algorithm for performing type checks on bespokes that requires only a single branch for all nodes in the current bespoke hierarchy. The new algorithm loads the bespoke index, performs an XOR, performs an AND, performs a CMP, then branches on the unsigned <= condition. The AND and CMP are elided when they are unnecessary, and the XOR is reduced in strength to a CMP when possible.

The algorithm to identify the proper values with which to XOR, AND, and CMP proceeds in O(N) time as follows:
1. Initialize the XOR value to 0, the AND value to 0xffff, and the CMP value to 0.
2. We consider the 0-th most significant bit.
3. If all of the live indices agree on this bit, set this bit in the XOR value to the value on which they agree.
4. Otherwise, if all of the dead indices agree on this bit, set this bit in the XOR value to the negation of the value on which they agree and set this bit in the CMP value to 1.
5. Otherwise, unset this bit in the AND mask to exclude it.
6. Masking off all bits that have not yet been considered, remove any dead indices that must fail the test and any live indices that must pass the test based on the current XOR/AND/CMP values.
7. Loop considering the next most significant bit.

This requires (log N) iterations. Computing agreement requires time O(M) where M is the size of the live and dead sets. The maximum size of the live and dead sets is O(N). Note that if conditions (3) or (4) are met, the maximum size of the live and dead sets is reduced by half. If condition (5) is met, then the live and dead sets are unchanged and we can reuse the result from the previous round. As a result, the overall runtime is O(N).

Strength reductions are as follows:
1. AND with 0xffff is removed.
2. CMP with 0 is removed.
3. AND with 0xfff and CMP with 0 relaxes the XOR to a CMP.

Per kshaunak's recommendation, I have retained the original single-mask algorithm to catch a few cases where we can save one instruction.

Differential Revision: D25250069

fbshipit-source-id: 7cf6a4b4431b4a9f87dda2d6b5f6641ebb579226