Simplify single bit type tests
Summary:
When a single bit test suffices for a bespoke layout check, use testi directly. vasm-simplify does not handle this case (and it is tricky to do so) because of the non-zero XOR before the AND.

To see why these operations suffice, note that we perform the computation `(xor(v, x) & a) == vx'a + v'xa`. If `x'a = 0`, this is equal to `v'xa`. When `xa` has a popcount of 1, then this is a test of a single bit of `v` and Z/NZ conditions negate to each other.

An analogous simplification could be performed for `xa = 0`, but this condition never arises in our algorithm.

Differential Revision: D25261322

fbshipit-source-id: af9582a2e50cb58615f8a2b0e478819685700a32