Fixup HHVM tests that include goto
Summary:
Aiming to remove support for `goto` and goto labels in D23328595.

Some of the HHVM test suite uses `goto` and goto labels. This diff removes those tests or modifies them to not use `goto`.

Separated out this removal to get an audit from HHVM to determine which tests should not be removed and instead should just be modified as the main purpose of those tests was not to test `goto`.

Reviewed By: jano

Differential Revision: D23352262

fbshipit-source-id: 2398d97ec360d939acef9340b2c6c230f080699c