Add dynamic_meth_caller
Summary:
We need a dynamic version of `meth_caller` to support migrating inst_meth and `$foo->$bar()` patterns in www. While we could use the existing `MethCallerHelper` wrapper for these dynamic `meth_caller`, the static version supports calling protected and private methods so long as the method is visible at the location that the `meth_caller` is constructed.

Other options I considered were (1) cloning or in some way evaling new `meth_caller` wrappers into existence to support dynamic `meth_caller`, (2) adding a new dynamic `meth_caller` type. This solution is far simpler (and safer to roll out even with the new native `meth_caller` still disabled). We add a new `DynMethCallerHelper` class which behaves identically to `MethCallerHelper` but expects a lambda which performs the actual dynamic call.

A change to hackc allows us to implicitly construct this lambda when we see calls to `dynamic_meth_caller`, and the lambda itself allows the method resolution to happen in the scope of the caller of `dynamic_meth_caller`. Note that we don't simply return a lambda as `DynMethCallerHelper` needs to expose a mechanism to query the class and method name (and potentially be memoizable).

Reviewed By: jano

Differential Revision: D25232333

fbshipit-source-id: b82601165b4de6c9b2cdf89b0309021fc92f4ce5