Rework ArrayLayout iterator APIs slightly
Summary: We drop BespokeIterAdvancePos - it's never advantageous for us to JIT code that calls a native helper to advance an iterator position - and add BespokeIterEnd. BespokeIterEnd can be efficiently implemented for all vecs and varrays, and for all monotype dicts. Rather than adding the specialized implementation to the MonotypeDict layouts, we "simply" make it a simplify pass - since the two ops have identical interfaces, we can reduce to the specialized op at any point at which we get type information.

Differential Revision: D25275170

fbshipit-source-id: bf1d41a8261147ed1827444cb8c18483d3fde4d1