Add a runtime option to control the low arena min address
Summary:
This allows the low arena started to be moved, e.g. to give more space to the translation cache. This diff only adds the option, but still leaves the low arena start at 1GB.

One small complication is that this creates a dependence from the util alloc library on the runtime library for the runtime options mechanism. This diff works around that by wrapping the read of the runtime option in an extern function declared in the runtime, while declaring a weak version of that function in util for tests and other binaries that link with this library.

Reviewed By: binliu19

Differential Revision: D25191337

fbshipit-source-id: 4083299e9aefb926cfa31a7101e38eb08e09bced